% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmvc.R
\name{corrmvc}
\alias{corrmvc}
\title{Correlation from a 2d covariance matrix.}
\usage{
corrmvc(mat)
}
\arguments{
\item{mat}{is a 2x2 covariace matrix}
}
\value{
The empirical correlation fo the two series
}
\description{
Computes the correlation given a covariance matrix of a bivariate variable.
}
