% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emp_agr}
\alias{emp_agr}
\title{Employment in agriculture}
\format{A ts vector a length of 25 observations from 1991 to 2015.}
\source{
\url{https://databank.worldbank.org/source/jobs#}
}
\usage{
emp_agr
}
\description{
Dataset of the % of total employment in agriculture (modeled ILO estimate) in OCDE members.
}
\keyword{datasets}
