% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_trend.R
\name{graph_trend}
\alias{graph_trend}
\title{Plot of original and smoothed time series.}
\usage{
graph_trend(dat, N, tau, dvar, label = NULL, jump = 1:N,
  bands = TRUE, las, bivariate = TRUE)
}
\arguments{
\item{dat}{is a 2x2 covariace matrix}

\item{N}{the number of observations}

\item{tau}{the smoothed time series}

\item{dvar}{the estimated variance for tau}

\item{label}{the vectors of caracters associated to the time points to appear in the axis}

\item{jump}{if label is too long, jump thin them on the axis}

\item{bands}{is TRUE to draw the approximately 95\%  confidence bands around tau}

\item{las}{is 1 and 2 if the asis should appear vertical and horizontal, respectively}

\item{bivariate}{is FALSE if dat is a univariate time series}
}
\value{
The empirical correlation fo the two series
}
\description{
It plots the univariate or bivariate. This function is not intended for users but to be called by trend_estimate.
}
