% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positive_definite.R
\name{positive_definite}
\alias{positive_definite}
\title{Checks if a squared matrix is positive definite and turn it to positive definied if necessary}
\usage{
positive_definite(m, c = NULL)
}
\arguments{
\item{m}{Is a 2x2 matrix.}

\item{c}{Is a small nonegative number.}
}
\value{
The same matrix (if positive definite) or its modification that is positive definite.
}
\description{
Checks if a squared matrix is positive definite and turn it to positive definied if necessary
}
