% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preliminar.R
\name{preliminar}
\alias{preliminar}
\title{Preliminar smoothing}
\usage{
preliminar(dat)
}
\arguments{
\item{dat}{is a 2-column matrix with the observations of a bivariate time series. Each row correspond to the values at a given time.}
}
\value{
The preliminary smoothen series ptau.

The final estimate for sigma.eta

The time series correlation given by sigma.eta, rho.eta.

The preliminary estimation for sigma_epsilon, sigma.epsilon,

A suggested value for lambda given by the empirical estimations.

The empirical time series correlation (preliminar to rho.eta), emp_rho.

The time series length N.
}
\description{
Obtains the preliminar smoothed series, based on the preliminar lambda value and empirical estimates for Sigma_eta and Sigma_epsilon. This function is called by trend_estimate as part of the smoothing process.
}
