% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psigma_estimates.R
\name{psigma_estimates}
\alias{psigma_estimates}
\title{Preliminar estimates}
\usage{
psigma_estimates(dat)
}
\arguments{
\item{dat}{is a 2 column matrix with for the bivariate time series observations. Each column correspond to the values at a given time.}
}
\value{
Sigma epsilon

Sigma eta
}
\description{
It computes the preliminary estimates of Sigma_epsilon and Sigma_eta
}
