% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigma_zf.R
\name{sigma_zf}
\alias{sigma_zf}
\title{Empirical cross-covarinace.}
\usage{
sigma_zf(h, vec1, vec2, N)
}
\arguments{
\item{h}{the lag value.}

\item{vec1}{observations for the first variable of the bivariate time series.}

\item{vec2}{observations for the second variable of the bivariate time series.}

\item{N}{the common length of vec1 and vec2.}
}
\value{
The value of lambda that corresponds to a smoothing level s.
}
\description{
Function that calculates the empirical cross-covariance of order h for a bivariate time series.
}
