% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_level.R
\name{smoothing_level}
\alias{smoothing_level}
\title{Smoothing value}
\usage{
smoothing_level(lambda, rho, N)
}
\arguments{
\item{lambda}{a nonegative number.}

\item{rho}{the correlation of the time series.}

\item{N}{the length of the observations.}
}
\value{
S
}
\description{
Function that reports the smoothing level for a given value of lambda, N and rho (=0 if univariate).
}
