% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trade}
\alias{trade}
\title{Annual Trade for USA and Mexico}
\format{An object of class \code{matrix} with 49 rows and 2 columns.}
\source{
\url{https://databank.worldbank.org/source/world-development-indicators#}
}
\usage{
trade
}
\description{
A dataset (matrix) containing the annual trade (% of GDP)  for USA and Mexico, from 1969 to 2017
}
\keyword{datasets}
